further mod breakup and some clippy
