+++ services/travelmux/src/bin/travelmux-server/main.rs
@@ -50,7 +50,7 @@ async fn main() -> Result<()> {
             .app_data(web::Data::new(app_state.clone()))
             .service(api::v4::plan::get_plan)
             .service(api::v5::plan::get_plan)
-            .service(api::v5::plan::get_directions)
+            .service(api::v5::directions::get_directions)
             .service(api::health::get_ready)
             .service(api::health::get_alive)
     })
